pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-bridge"
version = "0.4.3"
description = "Dual-backend PBPK Model Context Protocol server for qualification, verification, and reporting"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "PBPK MCP Team" }
]
license = { text = "Apache-2.0" }
keywords = ["mcp", "pbpk", "ospsuite", "rxode2", "ngra", "fastapi"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Framework :: FastAPI",
  "License :: OSI Approved :: Apache Software License"
]
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.29,<1.0",
  "pydantic>=2.6,<3.0",
  "python-dotenv>=1.0,<2.0",
  "structlog>=24.1,<25.0",
  "langchain-core>=0.2.10,<0.3",
  "langchain-community>=0.2.6,<0.3",
  "langgraph>=0.0.50,<0.1",
  "python-jose[cryptography]>=3.3,<4.0",
  "httpx>=0.27,<0.29",
  "celery>=5.4,<6.0",
  "boto3>=1.34,<2.0",
  "redis>=5.0,<6.0",
  "prometheus-client>=0.20,<1.0",
  "jsonschema>=4.23,<5.0"
]

[project.optional-dependencies]
dev = [
  "build>=1.2,<2.0",
  "pytest>=8.2,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "httpx>=0.27,<0.28",
  "ruff>=0.6.4,<0.7.0",
  "black>=24.4,<25.0",
  "mypy>=1.10,<2.0",
  "botocore-stubs>=1.34,<2.0",
  "moto[server]>=5.0,<6.0",
  "fakeredis>=2.23,<3.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008", "UP006", "UP007", "UP035", "E501", "I001", "UP037", "F401", "E731", "UP012", "UP034", "B023", "F821", "F841", "B904"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning"
]
asyncio_mode = "auto"
markers = [
  "compliance: MCP protocol compliance harness",
  "e2e: End-to-end regression suite covering reference PBPK models",
  "hpc_stub: HPC submission stub regression suite",
  "perf: Performance regression guard executed when benchmark artefacts are available",
]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
strict_optional = true
packages = ["mcp_bridge"]
exclude = ["tests"]
[[tool.mypy.overrides]]
module = [
  "langchain_core.*",
  "langchain_community.*",
  "langgraph.*"
]
ignore_missing_imports = true
